+++ Gear/GUI/LogicProbe/LogicAnalog.cs
@@ -22,6 +22,7 @@
  */
 
 using Gear.EmulationCore;
+using System;
 using System.Drawing;
 
 namespace Gear.GUI.LogicProbe
@@ -29,26 +30,25 @@ namespace Gear.GUI.LogicProbe
     /// @brief Provides a analog channel for LogicView.
     class LogicAnalog : LogicRow
     {
+        /// <summary></summary>
         private readonly LogicDigital[] Channels;
 
+        /// <summary></summary>
         public override string Name
         {
             get
             {
                 string s = "[";
                 foreach (LogicDigital r in Channels)
-                {
                     s += r.Name + ",\n";
-                }
                 return s.Substring(0, s.Length - 2) + "]";
             }
         }
 
-        public override int Height
-        {
-            get { return 64; }
-        }
+        /// <summary></summary>
+        public override int Height => 64;
 
+        /// <summary></summary>
         public override double MinTime
         {
             get
@@ -65,39 +65,52 @@ namespace Gear.GUI.LogicProbe
             }
         }
 
+        /// <summary>
+        /// Default constructor
+        /// </summary>
+        /// <param name="rows"></param>
         public LogicAnalog(LogicDigital[] rows)
         {
             Channels = rows;
         }
 
-        public override void Click()
-        {
-        }
-
+        /// <summary>
+        /// 
+        /// </summary>
+        public override void Click() { }
 
         /// @brief Clear samples when reset is needed.
-        /// 
         public override void Reset()
         {
-            foreach (LogicDigital r in Channels)
-            {
-                r.Reset();
-            }
+            foreach (LogicDigital channel in Channels)
+                channel.Reset();
         }
 
-        public override int Draw(System.Drawing.Graphics g, int top, float left, float width, double minTime, double scale)
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="graph"></param>
+        /// <param name="top"></param>
+        /// <param name="left"></param>
+        /// <param name="width"></param>
+        /// <param name="minTime"></param>
+        /// <param name="scale"></param>
+        /// <returns></returns>
+        /// @version v22.03.02 - Bugfix on comparisons of
+        /// <c>_channels[i].GetTime(index[i]) >= time</c>
+        public override int Draw(Graphics graph, int top, float left, float width,
+            double minTime, double scale)
         {
-            float height = Height * 3 / 4;
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
+            float height = Height * 3f / 4f;
             float bottom = top + Height;
 
             float previousX = width + left;
             float previousY = bottom;
             float nextX;
-            float nextY;
-            int[] index = new int[Channels.Length];
-
-            for (int i = 0; i < Channels.Length; i++)
-                index[i] = 0;
+            int[] index = new int[Channels.Length];  //automatically set to 0 all elements
 
             do
             {
@@ -105,13 +118,11 @@ namespace Gear.GUI.LogicProbe
                 int seekindex = 0;
 
                 for (int i = 1; i < index.Length; i++)
-                {
-                    if (Channels[i].GetTime(index[i]) > time)
+                    if (Channels[i].GetTime(index[i]) >= time)  //Corrected comparison for double
                     {
                         time = Channels[i].GetTime(index[i]);
                         seekindex = i;
                     }
-                }
 
                 int outputLevel = 0;
 
@@ -119,23 +130,21 @@ namespace Gear.GUI.LogicProbe
                     if (Channels[i].GetState(index[i]) == PinState.OUTPUT_HI)
                         outputLevel += 1 << i;
 
-                nextY = bottom - height * (outputLevel + 1) / (float)((1 << Channels.Length) + 2);
+                float nextY = bottom - height * (outputLevel + 1f) / ((1 << Channels.Length) + 2f);
                 nextX = (float)((time - minTime) / scale) * width + left;
 
                 if (nextX < left)
                     nextX = left;
 
-                g.DrawLine(Pens.Magenta, previousX, nextY, nextX, nextY);
-                g.DrawLine(Pens.Magenta, previousX, previousY, previousX, nextY);
+                graph.DrawLine(Pens.Magenta, previousX, nextY, nextX, nextY);
+                graph.DrawLine(Pens.Magenta, previousX, previousY, previousX, nextY);
 
                 previousX = nextX;
                 previousY = nextY;
 
                 for (int i = 0; i < index.Length; i++)
-                {
-                    if (Channels[i].GetTime(index[i]) == time)
+                    if (Channels[i].GetTime(index[i]) >= time)  //Corrected comparison for double
                         index[i]++;
-                }
 
                 if (Channels[seekindex].Overflow(index[seekindex]))
                     break;